[feat] 정보 등록 페이지 헤더 변경 및 상호작용 페이지로의 이동 구현
