+++ app/(tabs)/medicine/register/_layout.tsx
@@ -23,6 +23,18 @@ export default function RegisterLayout() {
           ),
         }}
       />
+      <Stack.Screen
+        name="registerForm"
+        options={{
+          headerShown: true,
+          title: "약물 등록",
+          headerLeft: () => (
+            <TouchableOpacity onPress={() => router.back()}>
+              <Ionicons name="chevron-back" size={24} color={"black"} />
+            </TouchableOpacity>
+          ),
+        }}
+      />
     </Stack>
   );
 }
