+++ app/(tabs)/medicine/register/registerForm.tsx
@@ -1,21 +1,107 @@
-import RegisterInfo from "@/components/register/RegisterInfo";
-import { Medication } from "@/types/medication";
+import Button from "@/components/Button";
+import Alarm from "@/components/register/Alarm/Alarm";
+import DateRange from "@/components/register/DateRange";
+import PerAOnce from "@/components/register/PerAOnce";
+import TakingCycle from "@/components/register/TakingCycle/TakingCycle";
+import TakingCycleDetails from "@/components/register/TakingCycle/TakingCycleDetails";
+import Group from "@/components/register/group/Group";
+import { colors } from "@/constants";
+import { DEV_SELECTED_MEDICINE } from "@/data/mockMedicine";
+import { useMedicationForm } from "@/hooks/useMedicationForm";
+import {
+  Medication,
+  SelectedMedicineInfo,
+  TakingType,
+} from "@/types/medication";
+import { useRouter } from "expo-router";
 import React from "react";
-import { View } from "react-native";
+import { ScrollView, StyleSheet, Text, View } from "react-native";
 
-const registerForm = () => {
-  const selectedMedicine = { id: "123", name: "우루사정" };
+export interface registerFormProps {
+  selected: SelectedMedicineInfo;
+  onSubmit?: (medication: Medication) => void;
+}
+const registerForm = ({
+  selected = DEV_SELECTED_MEDICINE,
+  onSubmit,
+}: registerFormProps) => {
+  const { medication, errors, updateField, validateForm } =
+    useMedicationForm(selected);
 
-  //테스트
-  const handleSubmit = (medication: Medication) => {
-    console.log("제출된 약물 정보: ", medication);
+  const handleSubmit = () => {
+    const { isValid } = validateForm();
+    if (isValid && onSubmit) {
+      onSubmit(medication);
+      useRouter().push("/medicine/register/interactionCheck");
+    }
+  };
+
+  const handleTakingTypeChange = (takingType: TakingType) => {
+    updateField("takingType", takingType);
+  };
+
+  const handleIntervalChange = (interval: number) => {
+    updateField("interval", interval);
+  };
+
+  const handleParticularDateChange = (dates: string[]) => {
+    updateField("particularDate", dates);
   };
 
   return (
-    <View style={{ flex: 1 }}>
-      <RegisterInfo selected={selectedMedicine} onSubmit={handleSubmit} />
-    </View>
+    <ScrollView style={styles.container} showsVerticalScrollIndicator={false}>
+      <View style={styles.nameContainer}>
+        <Text style={styles.name}>{selected.name}</Text>
+      </View>
+      <View>
+        <DateRange
+          startAt={medication.startAt}
+          endAt={medication.endAt}
+          onStartChange={(date) => {
+            updateField("startAt", date);
+          }}
+          onEndChange={(date) => {
+            updateField("endAt", date);
+          }}
+          errors={[...(errors.startAt || []), ...(errors.endAt || [])]}
+        />
+        <TakingCycle
+          selectedType={medication.takingType}
+          onTypeChange={handleTakingTypeChange}
+        />
+        <TakingCycleDetails
+          takingType={medication.takingType}
+          interval={medication.interval}
+          particularDate={medication.particularDate}
+          onIntervalChange={handleIntervalChange}
+          onParticularDateChange={handleParticularDateChange}
+        />
+        <PerAOnce />
+        <Alarm />
+        <Group
+          groupName={medication.groupName}
+          onGroupChange={(name) => updateField("groupName", name)}
+        />
+        <Button text="다음" onPress={handleSubmit} />
+      </View>
+    </ScrollView>
   );
 };
 
 export default registerForm;
+
+const styles = StyleSheet.create({
+  container: {
+    marginVertical: 25,
+    marginHorizontal: 10,
+  },
+  nameContainer: {
+    borderBottomWidth: 0.5,
+    borderBottomColor: colors.TEXT_GRAY,
+  },
+  name: {
+    fontSize: 23,
+    fontWeight: "bold",
+    paddingBottom: 10,
+  },
+});
